Created RadioComponent and shifted Modals
