+++ src/features/public/Donations/components/treeDonation/SelectCurrencyModal.tsx
@@ -3,53 +3,44 @@ import Fade from '@material-ui/core/Fade';
 import FormControl from '@material-ui/core/FormControl';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Modal from '@material-ui/core/Modal';
-import Radio from '@material-ui/core/Radio';
 import RadioGroup from '@material-ui/core/RadioGroup';
-import { withStyles } from '@material-ui/core/styles';
-import React, { useState } from 'react';
-import { getCountryDataBy } from '../../../../utils/countryUtils';
-let styles = require('./../styles/SelectCurrencyModal.module.scss');
+import React, { useEffect, useState } from 'react';
+import countriesData from '../../../../../utils/countriesData.json';
+import { sortCountriesData } from '../../../../../utils/countryUtils';
+import GreenRadio from '../../../../common/InputTypes/GreenRadio';
+let styles = require('./../../styles/SelectCurrencyModal.module.scss');
 
 export default function TransitionsModal(props: any) {
   const {
     openModal,
     handleModalClose,
-    taxDeductionCountries,
-    setCountry,
-    country,
     setCurrency,
+    setCountry,
     currency,
+    country,
   } = props;
 
-  const [countriesData, setCountriesData] = useState([]);
-  const [selectedModalValue, setSelectedModalValue] = useState(
-    `${country},${currency}`
-  );
+  const [sortedCountriesData, setSortedCountriesData] = useState(countriesData);
+  //   const [selectedModalCurrency, setSelectedModalCurrency] = useState(currency)
+  const [selectedModalValue, setSelectedModalValue] = useState(`${currency}`);
 
   // changes the currency in when a currency is selected
-  const handleCountryChange = (event: any) => {
+  const handleCurrencyChange = (event: any) => {
     setSelectedModalValue(event.target.value);
   };
 
   // changes the language and currency code in footer state and local storage
   // when user clicks on OK
   function handleOKClick() {
-    const selectedData = selectedModalValue.split(',');
-    setCountry(selectedData[0]);
-    setCurrency(selectedData[1]);
+    const selectedInfo = selectedModalValue.split(',');
+    setCurrency(selectedInfo[1]);
+    setCountry(selectedInfo[0]);
     handleModalClose();
   }
 
-  React.useEffect(() => {
-    setSelectedModalValue(`${country},${currency}`);
-  }, [country, currency]);
-
-  React.useEffect(() => {
-    const tempCountriesData: any = [];
-    taxDeductionCountries.forEach((countryCode) =>
-      tempCountriesData.push(getCountryDataBy('countryCode', countryCode))
-    );
-    setCountriesData(tempCountriesData);
+  // sorts the country data by country name as soon as the page loads
+  useEffect(() => {
+    setSortedCountriesData(sortCountriesData('countryName'));
   }, []);
 
   return (
@@ -69,13 +60,13 @@ export default function TransitionsModal(props: any) {
         <Fade in={openModal}>
           <div className={styles.modal}>
             <div className={styles.radioButtonsContainer}>
-              <p className={styles.sectionHead}>Select Country</p>
-              {/* maps the radio button for country */}
-              <MapCountry
-                countriesData={countriesData}
-                // this is selectedValue,
+              <p className={styles.sectionHead}>Select your Currency</p>
+              {/* maps the radio button for currency */}
+              <MapCurrency
+                sortedCountriesData={sortedCountriesData}
+                // this is selectedValue, country wala object
                 value={selectedModalValue}
-                handleChange={handleCountryChange}
+                handleChange={handleCurrencyChange}
               />
             </div>
 
@@ -98,8 +89,8 @@ export default function TransitionsModal(props: any) {
 }
 
 // Maps the radio buttons for currency
-function MapCountry(props: any) {
-  const { countriesData, value, handleChange } = props;
+function MapCurrency(props: any) {
+  const { sortedCountriesData, value, handleChange } = props;
   return (
     <FormControl component="fieldset">
       <RadioGroup
@@ -107,27 +98,16 @@ function MapCountry(props: any) {
         name="language"
         value={value}
         onChange={handleChange}
+        className={styles.currencyGrid}
       >
-        {countriesData.map((country: any) => (
+        {sortedCountriesData.map((country: any) => (
           <FormControlLabel
             value={`${country.countryCode},${country.currencyCode}`} // need both info
             control={<GreenRadio />}
-            label={`${country.countryName} · (${country.countryCode})`}
-            // label={`${country.countryCode},${country.currencyCode}`}
+            label={`${country.countryName} · ${country.currencyCode}`}
           />
         ))}
       </RadioGroup>
     </FormControl>
   );
 }
-
-// styles the radio button to green color for selected state
-const GreenRadio = withStyles({
-  root: {
-    color: '#000000',
-    '&$checked': {
-      color: '#89B53A',
-    },
-  },
-  checked: {},
-})((props) => <Radio color="default" {...props} />);
